scrapelight-backend/ml_worker/requirements.txt
celery==5.3.4
redis==5.0.1
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1
tqdm==4.66.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic
pydantic-settings
pydantic[email]
jinja2